refactor: 파일 구조 분리(LocationManager)
