+++ AppProduct/AppProduct/Core/Manager/Location/LocationManager.swift
@@ -8,29 +8,6 @@
 import Foundation
 import CoreLocation
 
-enum GeofenceEvent: Equatable {
-    case entered(String)
-    case exited(String)
-    case failed(String)
-}
-
-enum LocationError: LocalizedError {
-    case notAuthorized
-    case locationFailed(String)
-    case timeout
-    
-    var errorDescription: String? {
-        switch self {
-        case .notAuthorized:
-            return "위치 권한이 필요합니다."
-        case .locationFailed(let message):
-            return "위치를 가져올 수 없습니다. \(message)"
-        case .timeout:
-            return "위치 요청 시간이 초과되었습니다."
-        }
-    }
-}
-
 final class LocationManager: NSObject {
 
     // MARK: - Property
